Simplify startup to stop using isolation
