+++ tap/Program.cs
@@ -5,59 +5,51 @@
 using System;
 using System.Globalization;
 using System.IO;
-using System.Reflection;
+using OpenTap;
+using OpenTap.Cli;
+using System.Linq;
 
-namespace tap
+namespace tap;
+
+class Program
 {
-    class Program
+    private static bool IsColor()
     {
-        static void Main(string[] args)
-        {
-            CultureInfo.CurrentCulture = CultureInfo.InvariantCulture;
-            CultureInfo.DefaultThreadCurrentCulture = CultureInfo.InvariantCulture;
-            
-            // OPENTAP_INIT_DIRECTORY: Executing assembly is null when running with 'dotnet tap.dll' hence the following environment variable can be used.
-            Environment.SetEnvironmentVariable("OPENTAP_INIT_DIRECTORY", Path.GetDirectoryName(typeof(Program).Assembly.Location));
-            // in case TPM needs to update Tap.Cli.dll, we load it from memory to not keep the file in use
-            Assembly asm = null;
-            string entrypoint = "OpenTap.Cli.TapEntry";
-            try
-            {
-                string appDir = Path.GetDirectoryName(Assembly.GetEntryAssembly().Location);
-
-                Assembly load(string file)
-                {
-                    file = Path.Combine(appDir, file);
-                    if (File.Exists(file) == false)
-                        return null;
-                    return Assembly.Load(File.ReadAllBytes(file));
-                }
-                asm = load("Packages/OpenTAP/OpenTap.Cli.dll") ?? load("OpenTap.Cli.dll") ?? load("OpenTap.Cli.exe");
-                if (asm == null && File.Exists(Path.Combine(appDir, ".tapentry")))
-                {
-                    string[] lines = File.ReadAllLines(Path.Combine(appDir, ".tapentry"));
-                    asm = Assembly.Load(File.ReadAllBytes(Path.Combine(appDir, lines[0])));
-                    if (lines.Length > 1)
-                        entrypoint = lines[1];
-                }
-            }
-            catch
-            {
-                Console.WriteLine("Error finding OpenTAP CLI. Please try reinstalling OpenTAP.");
-                Environment.ExitCode = 7;
-                return;
-            }
-            if (asm == null)
-            {
-                Console.WriteLine("Missing OpenTAP CLI. Please try reinstalling OpenTAP.");
-                Environment.ExitCode = 8;
-                return;
-            }
-
-            var type = asm.GetType(entrypoint);
-            var method = type.GetMethod("Go", BindingFlags.Static | BindingFlags.Public);
-            method.Invoke(null, Array.Empty<object>());
-        }
+        string[] arguments = Environment.GetCommandLineArgs();
+        if (arguments.Contains("--color") || arguments.Contains("-c"))
+            return true;
+        var envvar = Environment.GetEnvironmentVariable("OPENTAP_COLOR");
+        if (envvar == null)
+            return false;
+        if (envvar == "always")
+            return true;
+        else if (envvar == "auto")
+            return !(Console.IsErrorRedirected || Console.IsOutputRedirected);
+        else if (envvar != "never")
+            Console.WriteLine("Unknown value of variable OPENTAP_COLOR, valid values are always, auto or never.");
+        return false;
     }
+        
+    static void Main(string[] args)
+    {
+        CultureInfo.CurrentCulture = CultureInfo.InvariantCulture;
+        CultureInfo.DefaultThreadCurrentCulture = CultureInfo.InvariantCulture;
+            
+        // OPENTAP_INIT_DIRECTORY: Executing assembly is null when running with 'dotnet tap.dll' hence the following environment variable can be used.
+        Environment.SetEnvironmentVariable("OPENTAP_INIT_DIRECTORY", Path.GetDirectoryName(typeof(Program).Assembly.Location));
+            
+        var start = DateTime.Now;
+        ConsoleTraceListener.SetStartupTime(start);
 
+        bool isVerbose = args.Contains("--verbose") || args.Contains("-v");
+        bool isQuiet = args.Contains("--quiet") || args.Contains("-q"); ;
+        bool isColor = IsColor();
+        var cliTraceListener = new ConsoleTraceListener(isVerbose, isQuiet, isColor);
+        Log.AddListener(cliTraceListener);
+        AppDomain.CurrentDomain.ProcessExit += (s, e) => cliTraceListener.Flush();
+
+        PluginManager.Search();
+        DebuggerAttacher.TryAttach();
+        CliActionExecutor.Execute();
+    }
 }
